Improved the deployment script. Added some documentation to the ReadMe

+++ dev-bin/deploy.sh
@@ -1,6 +1,16 @@
 #!/bin/sh
 
+function silentSsh {
+    local connectionString="$1"
+    local commands="$2"
+    if [ -z "$commands" ]; then
+        commands=`cat`
+    fi
+    ssh -T $connectionString "$commands"
+}
+
 SPECIFIED_VERSION="$1"
+REMOTE_CONNECTION="$2"
 
 clear
 
@@ -10,22 +20,38 @@ then
   exit 1;
 fi
 
+if [ -z "$REMOTE_CONNECTION" ]
+then
+  echo "You must specify a connection string"
+  exit 1;
+fi
+
 echo "Deploying OpenMud Version: $SPECIFIED_VERSION"
 
 echo "Copying tarball to server"
-scp "builds/openmud-$SPECIFIED_VERSION.tar.gz" matt@192.168.0.44:/deploy/openmud/releases
+scp "builds/openmud-$SPECIFIED_VERSION.tar.gz" "$REMOTE_CONNECTION:/deploy/openmud/releases"
 
-ssh -T matt@192.168.0.44 <<EOF
+silentSsh "$REMOTE_CONNECTION" <<EOC
   cd /deploy/openmud/releases
-  mkdir openmud-$SPECIFIED_VERSION
-  cd openmud-$SPECIFIED_VERSION
-  mv ../openmud-$SPECIFIED_VERSION.tar.gz .
-  tar -xzf openmud-$SPECIFIED_VERSION.tar.gz
-  npm i --production
+
+  echo "Creating release location"
+  mkdir "openmud-$SPECIFIED_VERSION"
+  cd "openmud-$SPECIFIED_VERSION"
+  mv "../openmud-$SPECIFIED_VERSION.tar.gz" .
+
+  echo "Extracting tarball"
+  tar -xzf "openmud-$SPECIFIED_VERSION.tar.gz"
+
+  echo "Installing dependancies"
+  npm i -s --production
   cd ../../
+
+  echo "Updating 'current' symlink"
   rm -f current
-  ln -s releases/openmud-$SPECIFIED_VERSION current
+  ln -s "releases/openmud-$SPECIFIED_VERSION" current
   cd current
   cd ../
-  pm2 startOrRestart current/ecosystem.json
-EOF
+
+  echo "Restarting PM2 process"
+  pm2 -s startOrRestart current/ecosystem.json
+EOC
